fix: move url builders to common
