+++ apps/ui/modules/apolloClient.ts
@@ -5,10 +5,11 @@ import { onError } from "@apollo/client/link/error";
 import { GraphQLError } from "graphql";
 // eslint-disable-next-line unicorn/prefer-node-protocol -- node:querystring breaks the app
 import qs, { ParsedUrlQuery } from "querystring";
-import { isBrowser } from "./const";
 import { GetServerSidePropsContext, PreviewData } from "next";
 import { IncomingHttpHeaders } from "node:http";
+import { buildGraphQLUrl } from "common/src/urlBuilder";
 import { env } from "@/env.mjs";
+import { isBrowser } from "./const";
 
 const errorLink = onError(({ graphQLErrors, networkError }) => {
   if (graphQLErrors) {
@@ -55,15 +56,7 @@ export function createApolloClient(
     ? getServerCrsfToken(ctx?.req?.headers)
     : getCookie("csrftoken");
 
-  // a hack to workaround node-fetch dns problems with localhost
-  // this will not work with authentication so when we add authentication to the SSR we need to fix it properly
-  // TODO add an url builder (with the fetch hack boolean)
-  const shouldModifyLocalhost =
-    env.ENABLE_FETCH_HACK && !isBrowser && hostUrl.includes("localhost");
-  const apiUrl = shouldModifyLocalhost
-    ? hostUrl.replace("localhost", "127.0.0.1")
-    : hostUrl;
-  const uri = `${apiUrl}${apiUrl.endsWith("/") ? "" : "/"}graphql/`;
+  const uri = buildGraphQLUrl(hostUrl, env.ENABLE_FETCH_HACK);
 
   const enchancedFetch = (url: RequestInfo | URL, init?: RequestInit) =>
     fetch(url, {
